Coverage cap now handles fraction formats
If program coverage is a fraction, it does not need to be compared
against a number of people.
The coverage value itself will reflect excess if it is greater than 1.
Thus fraction-format coverages can now be provided to programs in the
databook and still be capped appropriately.
Note: The maximum value attainable for a parameter under the
coverage-capping process is coverage times impact factor, so the impact
factor should be less than 1 in general.