Set up parset selection for budget scenario
Started the budget scenario GUI by putting in two combo boxes for parset
and progset selection.
The first one is implemented and is likely the only one that will matter
(if the user wants to try various calibrations), as programs will be
loaded in as a default progset.
However, proceeding to set up the second combo box.