Handle FigureCanvasQTAgg versions
As matplotlib is separate from PyQt, it will have no problem loading
qt4agg or qt5agg, regardless of the PyQt version installed on a user's
system.
Therefore, the qt-agg imports should also be part of the try-except
clause in importPyQt().
This commit should import the corresponding version correctly and return
the import to the global namespace (unlike a previous attempt where the
import was scoped locally).