Start year works
It is possible to run budget scenarios for the default options
dictionary and change, via widget, the starting year for budget-based
overwrite.
If values are entered that cannot be converted into a float, then the
GUI complains and reverts the start year widget to the value for the
last options dictionary that worked.
Next up; allowing the init_alloc values to be modified.