Budget Scenario GUI appears to work
Whenever a ProgramSet is loaded (or a budget scenario fails due to poor
options being provided by the user), the budget scenario GUI generates a
list of program/budget widgets, filled in with default values.
The user can vary these values however they wish and run the scenario,
generating associated results and plotting them in the same way as done
for the calibration GUI.
Thus, as far as can be seen, the budget scenario GUI seems to work as
intended and, short of bugs, is complete.