Coverage cap overflow fix
Coverages for certain programs can be precalculated as a time-dependent
array in Model.prog_vals rather than a time-independent value.
This means that overflow_list would occasionally arise as a list of
lists and cause problems.
This commit thus introduces a try-except clause that extracts the right
coverage value related to the correct timepoint and hopefully makes
overflow work appropriately.