Coverage scenarios do not crash with fixed costs
This commit returns a budget of NaN for a fixed-cost program that is
given an allocation in the form of a budget, rather than throwing an
exception.
The special NaN value should identify potential future bugs based on
fixed-cost programs with specified coverages, while still allowing
coverage scenarios with fixed-cost programs in the allocation to
proceed.
After all, a program's budget is only used in model.py if the program
has an impact parameter, which fixed-cost programs do not (at least as
intended).