Autocalibrate flags added to parameter
First up, the cascade xlsx files now head 'Calibrate?' columns as
'Autocalibrate'.
Characteristics and parameters can be flagged as 'n' or '-1' to specify
that they are not to be autocalibrated, i.e. that y_factor in settings
is -1, while any other flag will give y_factor a default value of 1.
Now, in an attempt to make minimal changes, settings/cascade/data
operate as usual.
However, during data loading into parameters, any y_factors marked -1
are converted to 1. The only difference is that the ones marked -1
involve a new Parameter.autocalibrate attribute to be marked as False,
while any other values mark it as True.
The effect of this is yet to be propagated to autocalibration, but all
models should operate as normal provided that they have the right
'Autocalibrate' heading.