Updating calibration for new ASD
The modern version of ASD uses 'maxtime' and 'maxiters' as arguments, so
code has been changed throughout the codebase to reflect this.
If this transition is completed, asd_autocalibrate.py can be deleted.
Unfortunately, ParameterSet.update() does not take well to subsets of
calibrateable parameters.
Modifications to this method are intended for the next commit.