Allow autocalibration to take temp parsets
All GUI parset operations are applied to a temporary parset that is
stored as an attribute of the GUI.
User messages may need to emphasise this, so that users understand this
parset is not autosaved.
In any case, Project.runAutofitCalibration() now takes in an optional
parset argument that is prioritised over old_parset_name.
This allows for GUI parset changes to actually take effect, such as
checking parameters/characteristics.