Streamlined calibration and updated ASD
Cascade parameters and characteristics are now both updated via the same
ParameterSet method, thanks to ParameterSet.getPar() being able to
select either type.
They can also be updated by an arbitrary subset of populations due to
propagation of population labels from calibration.py to the
ParameterSet.update() method.
Further code edits have aligned calibration.py with the latest version
of asd.py used in the HIV codebase.
It is now possible to autocalibrate via the GUI for any selection of
parameters and characteristics, provided that the fit is for all
characteristics and related data.