Exception catch for invalid parset change
The most predictable error of autocalibration is if an initial
characteristic size is changed so that it is smaller than its subsets or
greater than its supersets.
A try-except clause is thus implemented in objective calculation for
running invalid simulations; in this case, a fitting score of np.inf is
returned.