Minor edit for value deletion
Attempting to delete the contents of a GUI parset table element that did
not correspond to an actual stored parameter value would be interpreted
as a failure to delete due to 'only one value being in the parameter
array', even if there were many others, and result in interpolation.
This behaviour has been nixed; deleting an empty cell properly reports a
successful deletion.
Note: An attempt to disable characteristic elements before the first
year resulted in success, but table-generation performance was too slow
to retain it. Also, the autocalibrate version of asd has been discarded
from the codebase now that there are no references to it.