Population-based fitting metric
The calibration.py file has been upgraded to allow for population
exclusion to affect the fitting calculation, albeit in a hacky manner.
The remaining problem to solve is that not all characteristics in
project data make it to the parameter set for some reason.
With Peru, in particular, this means that deselecting all
characteristics in the parset GUI still results in certain latency
characteristics being used in fitting metric calculations.
The rationale for the exclusion must be re-examined.