Calibration t-vector fix
When calculating the fitting metric, the calibration framework compares
data points against corresponding result values.
However, the original version of autocalibration would attempt to
calculateFitFunc() by passing in results.t_observed_data, a
yearly-spaced time-vector.
In effect, the indices extracting corresponding result values would not
account for timestep, e.g. a 2010 value would be considered to be at the
11th timestep, when the 2002.5 value would actually be extracted.
Fortunately, the Result object already contains the full dt-based time
vector as 't_step'. This was propagated instead and now the fitting
metric works properly.
This means that autocalibration seems to be a completely successful
process.