Get tagged transitions plotting in GUI again
Via if/elif clauses and nexted try/except branches, the GUI has been
updated to fill up a combo box with characteristic and parameter names,
the latter only if a parameter has a transition tag and/or is marked as
an output.
During plotting, transition tags take precedence and are used as a proxy
label for a parameter, so the getValuesAt() return effective flow rates.
Otherwise, a parameter label simply marks that a parameter should be
plotted as a calculated output of a model run.
In essence, GUI users are now able to plot custom parameters without
neglecting effective transition rates.