Carry over budget scenario basics for reconciliation
A lot of the basic infrastructure for program reconciliation is similar
to the basics used in running budget scenarios.
This is unsurprising as both involve loading projects and then running
them for certain parsets and progsets.
Ideally, the budget scenario GUI should inherit a base that is common
with program reconciliation, or even just be an extension on the same
window.
However, for expediency's sake, the two are being coded separately and
can be fused in the future.