Program set allows value insertion
A method has been constructed called Program.insertValuePair() that
mimics the Parameter version.
If a year value is entered as an argument that already exists within a
Program, the corresponding 'attribute' value is overwritten by something
user-provided.
If not, the 't' array has a new year appended to it, while all
'attribute' arrays are appended with a np.nan value, including
Program.cost and Program.cov.
The appropriate 'attribute' value, corresponding to the newly appended
year, is then overwritten by the user-provided value.
Naturally, the appending of values requires Program.interpolate() to
account for unordered attributes, which it currently does due to
sorting in interpolateFunc().