Manual budget update via GUI complete
Using the new Program value-pair insertion method, budget edits in the
progset table are now applied to the attached progset.
Whether this persistence is preferred, affecting interpolated values for
other start years, is up for debate.
Nonetheless, it is assumed that the user updates values intelligently.
Errors can of course be reverted; without saving the progset, it is easy
to refresh all budget values to databook versions by simply selecting
'default' progset from the combo box.
The next stage is to now connect up the program reconciliation code with
the GUI, modifying as required.