Enable min-max parameter constriction
The cascade now allows for parameters to be flagged with an unvalidated
'Minimum Value' and/or 'Maximum Value'.
These constrain the values of parameters during a model run, once when
Model.build() constructs links and deps out of a parset, and then at the
end of Model.updateValues() for functions.
The efficiency of this process needs to be reviewed, as well as its
validity. Simple test cases work, but extended validation would be
ideal.