Reconcile button activated in GUI
A reconciliation analogue to the autocalibration interface, i.e. for
editing time and activating the process with a button, has been enabled
in the GUI. It is currently not connected to the reconciliation process.
As a side note, error messages for autocalibration have been updated to
reference self.parset.name (i.e. the temporary ParameterSet name) rather
than self.parset_name (i.e. the name of a permanent parset stored in
self.progset).