Reconcile operates on progset with time limit
Project.reconcile() can now operate on distinct progsets rather than
requiring a name for reference.
However, due to the way that reconcileFunc() is set up, this progset
must be temporarily attached to a Project if it is to be used in the
current code.
A time limit has also been propagated from the GUI to reconcileFunc().