Reconcile function operates on a year
The reconciliation function code was originally updating the last
time-dependent value of Program attributes, which produced confusing
behaviour when viewed with a GUI that allows for reconciling other
years.
The updateProgset() function has thus been updated to utilise the new
Program.insertValuePair() method.
Tests are still underway to see if unit costs and budgets are being
modified by the reconciliation process, as activated by the GUI.