Sigma dict introduced for reconciliation
The createAttributeList() function in reconciliation.py has been heavily
modified to pass back, as a secondary return, an index_dict that maps
all indices of the primary return, attribute_list, to a 'prog_label' and
'type'.
This means that reconcileFunc() now uses the typing provided by
index_dict to determine what sigmas for the attribute_list should be.
Currently, this commit has been tested to maintain functionality for the
original kwarg 'sigmas'.
However, a new kwarg named 'sigma_dict' has been propagated out to
Project, mapping prog_label to three different sigmas keyed by 'budget',
'unit_cost' and 'attribute'.
This should allow greater generalisation during reconcileFunc(), as
provided by the GUI.
The final piece of the puzzle is to connect the checkboxes and
user-input sigmas in the GUI with the sigma_dict kwarg.