Sigma values and checkboxes editable
The reconciliation GUI now contains a sigma_dict attribute that
initialises whenever a progset is loaded and takes in a sigma value from
the reconciliation table if the corresponding box is ticked.
It is also possible to select or edit all checkboxes for a sigma type
simultaneously, although duplicating the selection and values acts in
concert.
The GUI will be edited in the next commit to make this clear.