Standard GUI reconciliation process complete
As of this commit, it is possible to manipulate unit costs and budgets
(for selected years) within a program set.
It is also possible to run an ASD-led reconciliation process between
parset and progset for the chosen year, with user-selected sigmas across
program unit costs, budgets and attributes.
Testing thus far has allowed for selective reconciliation (e.g. by only
ticking 0.2 for a DS treatment program).
Thus, this commit marks the conclusion of developing the GUI
reconciliation process.
From here on out, there are notes, as there is a lot that could/should
be improved regarding the reconciliation process in the future.
From the interface side, the checkboxes seem redundant if a sigma of 0
serves the same purpose; these can probably be taken out.
Also, attribute changes are currently very opaque, with no way for the
user to see how they have changed, let alone manually manipulate them.
Perhaps they could be attached to the output returned by
reconcileFunc().
Plenty of quality assurance is also recommended to make sure that the
reconciliation process itself is working as desired.