GUI allows program attributes to be updated
Why not? As the progset table already does the difficult work in pulling
attributes from programs and marking them in the table, it was not
difficult to allow edits to those values to carry across into the
progset.
Note: Simple tests showed some curious results. A sensitivity boost for
mass screening with South Africa ended up making the epidemic worse.
However, as this is an apps-related issue, with recovery and
reactivations boosted, it is just made as a note here.