Fixes for autocalibration process
This commit hopefully fixes a number of errors in the way that the
calibration GUI interacts with the autocalibration process.
First up, the vertical length of the calibration table now accounts for
characteristics, meaning that parameters are not cut off from the
user-interface (potentially being considered for calibration even if the
user has the impression that everything has been deselected).
Secondly, ParameterSet updating informs which parameter and population
are being modified for better debugging purposes.
Thirdly, characteristic extraction for the target of perturbation now
excludes those without entry points.
Finally, the usage of keys() has been slashed in gui.py for performance
reasons.
Thus, limited autocalibration tests on a PNG project now properly select
variants and targets, operating in a timely fashion.