GUI reconciliation fix
When updating values that are not in column 0 or 2 of the progset table,
progset attributes are edited.
This is a problem if column 4 is implicity (i.e. not by the user)
updated.
Therefore a suitable catch has been implemented in this commit to
restrict attribute updates to the attribute columns.