GUI budget updates take effect
The GUI progset table edits values in the progset, but, given that
runSim() options requires an init_alloc to be provided, simply editing
cost values was not enough.
This commit actually regenerates the defaultOptimOptions() whenever
running a simulation for the reconciliation GUI, thus updating
init_alloc.
The function defaultOptimOptions() also allows an optional kwarg in the
form of a different program start year.