Attempt to convert coverage with timestep
Program coverages are annual. It follows that the program coverages can
be distributed 1/dt times or less between timesteps of dt length.
Coverages should thus be converted from annual to quarterly before
comparing against the compartment popsize available to transfer.
The downside is that, for probabilistic transition functions, a coverage
conversion compatible with the timestep conversion formula does not
exist beyond saturation.
It is assumed that annual coverage overflow is equivalent to quarterly
coverage overflow beyond single-program saturation. This can be reviewed
later.