Modified post-saturation overflow value
For fraction-based parameters, annual coverage and its corresponding
overflow value has a maximum beyond which it cannot expand without
violating the laws of probability (relating to transition time-step
conversions).
The previous shift to a linear increase of overflow beyond this point
led to bizarre phenomena.
Fixing the overflow value beyond saturation, on the other hand, leads to
improved behaviour.