Restructure of program overwrites in model
Every step of impact calculation is restructured and made explicit.
Overflow corresponds to a dt-ly conversion of coverage.
The only functional difference over the previous model code is that
program impact is always capped at 1.0 to avoid transition conversion
issues.
This is usually rescaled down of course due to coverage capping.