Timestep based impact conversion
Previously, annual impact was not converted to dt-ly rates when annual
coverage was.
Because overflow is dt-ly coverage, this meant that overflow could be
less than annual impact.
This was not a problem when only one program was targeting a parameter;
if dt-ly coverage was below annual impact, the annual impact was already
below 100% and thus fine.
However, with multiple programs, the impacts add up, leading to
situations where net overflow is less than new (annual) impact and
unable to renormalise impact below 100%.
This commit inelegantly (in order to not disturb the implementation of
impact constraints and non-transition program targets) stores dt-ly
impacts for renormalisation, which are all smaller than corresponding
coverage overflows.
This means that the sum of renormalised impacts is always less than 1.
This does however mean an extra annualisation step is required following
overflow-based renormalisation, once again either linear or nonlinear
depending on target parameter format.