Reconciliation fixes
First up, both the model and reconciliation metric now only cap coverage
for individual programs if the parameter they are converting into is of
fraction format; there is no difficulty with rescaling unconstrained
coverage for number formats.
This will be properly tested once cascades are built with number-format
transitions.
Secondly, the reconciliation metric now accumulates an 'extra metric'
(currently set to net coverage) for programs that have fractional
impacts greater than 1 (with target parameters of fraction-format).
This extra metric acts as an overwhelmingly strong disincentive for
individual programs to go beyond saturation in these particular cases.
Finally, a major fix has been applied to the reconciliation process,
which, for the default option of budget constraint, would calculate
metrics for constrained budgets but would then return the unconstrained
version to the GUI; this unconstrained 'allocation' would not match up
with the optimal constrained 'allocation', leading to bizarre behaviour.
The constraint procedure is now applied after ASD is completed as well.
However, for GUI purposes (where users select one program budget to
modify with the understanding that others will not change) the GUI hook
in to project reconciliation now turns of the budget-constraint default.